Add ajax utils
This introduces a design pattern for fetching data from ajax requests
and consuming the result in a synchronous way (without callback
functions).

The implementation for the Legend widget has been generalized into ajax
utils that can be used for other components/plugins.

The purpose of this is to avoid deeply nested async callback trees (think
layer selector). Additionally, the cache is shared throughout the app so
plugins fetching the same data won't make extra ajax requests (Ex. if
you open the same plugin in Split View).