Add support for adjusting layer opacity.
Allows users to set the opacity on a layer by layer basis for those
layers that are part of services that support dynamic layers.

* Set individual layer opacity using the LayerDrawingOptions class. This
feature is only supported for layers that are served from a service with
support for dynamic layers turned on. This is not the default setting.
* Save layer opacity settings are part of the state.
* Adjust the layer menu to show or hide the opacity slider depending on
the layer's support for opacity.
* Add opacity to the schema. If this value is present for a layer in the
* config, the initial opacity of layer is set to this value.
* Add Alabama layers to the config to demo the opacity slider. This
service is currently the only one that supports dynamic layers.

Refs #522