Prevent events from being handled on both maps.
* Previously, two events were fired from listeners on the
layer menu. I believe this was because each layer selector
instance has it's own set of templates, and because of the
way the listener is set up, both instances of the layer
menu were being captured. To fix this, we give each
instance of the layer menu a unique id that we can bind
events to. We could have also created the layer selector
DOM element as a child of the layer selector, but this made
positioning the element difficult.

Refs #553