Add layer info box to state.
* The layer info box is triggered from the layer selector, but
is it's own DOM element (as well as an eye icon in the layer selector
that has an active state) that must persist through layer selection
and layer selector rerenderings. To make this easier, the state
of the layer info box is tracked in the state, as opposed to
manipulating the DOM after it is rendered.
* Add methods for getting, settings, clearing, and checking the
info box state.
* Rerender the layer selector when the info box is closed or open.
* Add the active class to the eye icon for the layer that an info
box open.

Refs #545