Move id logic back to LayerNode
I thought we needed to move this logic to Tree but that posed a problem
when testing the value of a LayerNode id outside of the scope of a Tree
structure. I came up with this solution which allows us to obtain the
value of a layer ID without having to fully bake the object first.